Refactors the get request
Inspired by Neight's API, we've implemented the get request using .then and .catch
we also seperated the FoodController and used the .Router() Express method to
do so. We implemented in the Food model a .find method that will
return undefined if the Knex SQL query fails. Tests passing, moving on to POST.